Pull CAD object code out of objects.c
CAD objects are a feature added to Xastir in support of search and
rescue management tasks, but have nothing whatsoever to do with APRS
objects.

This commit simply pulls all of the code and static data for handling
CAD objects out of objects.c and puts them into a new file,
cad_objects.c (and associated header, cad_objects.h).

Doing this makes objects.c smaller and will aid greatly in my work on
refactoring APRS object handling code.

There are no code changes in this commit.  All it does is move some
functions from one file to another, and make sure that all the files
that *reference* that code use the correct include file.
